compiler: Move extract_increments in to clusterize
