+++ devito/core/gpu.py
@@ -6,8 +6,8 @@ from devito.core.operator import CoreOperator, CustomOperator, ParTile
 from devito.exceptions import InvalidOperator
 from devito.passes.equations import collect_derivatives
 from devito.passes.clusters import (Lift, Streaming, Tasker, blocking, buffering,
-                                    cire, cse, extract_increments, factorize,
-                                    fission, fuse, optimize_pows)
+                                    cire, cse, factorize, fission, fuse,
+                                    optimize_pows)
 from devito.passes.iet import (DeviceOmpTarget, DeviceAccTarget, mpiize, hoist_prodders,
                                is_on_device, linearize, relax_incr_dimensions)
 from devito.tools import as_tuple, timed_pass
@@ -178,7 +178,6 @@ class DeviceAdvOperator(DeviceOperatorMixin, CoreOperator):
             clusters = blocking(clusters, sregistry, options)
 
         # Reduce flops
-        clusters = extract_increments(clusters, sregistry)
         clusters = cire(clusters, 'sops', sregistry, options, platform)
         clusters = factorize(clusters)
         clusters = optimize_pows(clusters)
