+++ devito/ir/clusters/algorithms.py
@@ -5,13 +5,14 @@ import numpy as np
 import sympy
 
 from devito.exceptions import InvalidOperator
-from devito.ir.support import Any, Backward, Forward, IterationSpace
+from devito.ir.support import Any, Backward, Forward, IterationSpace, PARALLEL_IF_ATOMIC
 from devito.ir.clusters.analysis import analyze
 from devito.ir.clusters.cluster import Cluster, ClusterGroup
-from devito.ir.clusters.queue import Queue, QueueStateful
+from devito.ir.clusters.visitors import Queue, QueueStateful, cluster_pass
 from devito.symbolics import uxreplace, xreplace_indices
 from devito.tools import (DefaultOrderedDict, Stamp, as_mapper, flatten, is_integer,
                           timed_pass)
+from devito.types import Symbol
 from devito.types.dimension import BOTTOM, ModuloDimension
 
 __all__ = ['clusterize']
@@ -35,6 +36,9 @@ def clusterize(exprs, options=None, **kwargs):
 
     # Determine relevant computational properties (e.g., parallelism)
     clusters = analyze(clusters, options)
+
+    # Input normalization (e.g., SSA)
+    clusters = normalize(clusters, **kwargs)
 
     return ClusterGroup(clusters)
 
@@ -314,3 +318,35 @@ class Stepper(Queue):
             processed.append(c.rebuild(exprs=exprs, ispace=ispace))
 
         return processed
+
+
+def normalize(clusters, **kwargs):
+    sregistry = kwargs['sregistry']
+
+    clusters = normalize_reductions(clusters, sregistry)
+
+    return clusters
+
+
+@cluster_pass(mode='all')
+def normalize_reductions(cluster, sregistry):
+    """
+    Extract the right-hand sides of reduction Eq's in to temporaries.
+    """
+    if not any(PARALLEL_IF_ATOMIC in v for v in cluster.properties.values()):
+        return cluster
+
+    processed = []
+    for e in cluster.exprs:
+        if e.is_Increment and e.lhs.function.is_Input:
+            handle = Symbol(name=sregistry.make_name(), dtype=e.dtype).indexify()
+            if e.rhs.is_Number or e.rhs.is_Symbol:
+                extracted = e.rhs
+            else:
+                extracted = e.rhs.func(*[i for i in e.rhs.args if i != e.lhs])
+            processed.extend([e.func(handle, extracted, operation=None),
+                              e.func(e.lhs, handle)])
+        else:
+            processed.append(e)
+
+    return cluster.rebuild(processed)
