+++ devito/passes/clusters/cse.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 
-from devito.ir import Cluster, Scope
-from devito.passes.clusters.utils import cluster_pass, makeit_ssa
+from devito.ir import Cluster, Scope, cluster_pass
+from devito.passes.clusters.utils import makeit_ssa
 from devito.symbolics import count, estimate_cost, q_xop, q_leaf, uxreplace
 from devito.types import Eq, Symbol
